Commonly used cardiovascular risk scores for the prediction of major cardiac events (MACE)

This function implements seven cardiovascular risk scores row wise in a data frame with the required variables. It would then retrieve a data frame with two extra columns for each risk score including their calculations and classifications

Usage

calc_scores(
  data,
  typical_symptoms.num_df = "typical_symptoms.num",
  ecg.normal_df = "ecg.normal",
  abn.repolarisation_df = "abn.repolarisation",
  ecg.st.depression_df = "ecg.st.depression",
  Age_df = "Age",
  diabetes_df = "diabetes",
  smoker_df = "smoker",
  hypertension_df = "hypertension",
  hyperlipidaemia_df = "hyperlipidaemia",
  family.history_df = "family.history",
  atherosclerotic.disease_df = "atherosclerotic.disease",
  presentation_hstni_df = "presentation_hstni",
  Gender_df = "Gender",
  sweating_df = "sweating",
  pain.radiation_df = "pain.radiation",
  pleuritic_df = "pleuritic",
  palpation_df = "palpation",
  ecg.twi_df = "ecg.twi",
  Δ2nd_hstni_df = "Δ2nd_hstni",
  killip.class_df = "killip.class",
  heart.rate_df = "heart.rate",
  systolic.bp_df = "systolic.bp",
  aspirin_df = "aspirin",
  number.of.episodes.24h_df = "number.of.episodes.24h",
  previous.pci_df = "previous.pci",
  previous.cabg_df = "previous.cabg",
  total.chol_df = "total.chol",
  total.hdl_df = "total.hdl",
  Ethnicity_df = "Ethnicity"
)

Arguments

typical_symptoms.num

a numeric vector of the number of typical symptoms

ecg.normal

a binary numeric vector, 1 = yes and 0 = no

abn.repolarisation

a binary numeric vector, 1 = yes and 0 = no

ecg.st.depression

a binary numeric vector, 1 = yes and 0 = no

Age

a numeric vector of age values, in years

diabetes

a binary numeric vector, 1 = yes and 0 = no

smoker

a binary numeric vector, 1 = yes and 0 = no

hypertension

a binary numeric vector, 1 = yes and 0 = no

hyperlipidaemia

a binary numeric vector, 1 = yes and 0 = no

family.history

a binary numeric vector, 1 = yes and 0 = no

atherosclerotic.disease

a binary numeric vector, 1 = yes and 0 = no

presentation_hstni

a continuous numeric vector of the troponin levels

Gender

a binary character vector of sex values. Categories should include only 'male' or 'female'.

sweating

a binary numeric vector, 1 = yes and 0 = no

pain.radiation

a binary numeric vector, 1 = yes and 0 = no

pleuritic

a binary numeric vector, 1 = yes and 0 = no

palpation

a binary numeric vector, 1 = yes and 0 = no

ecg.twi

a binary numeric vector, 1 = yes and 0 = no

2nd_hstni

a binary numeric vector, 1 = yes and 0 = no

killip.class

a numeric vector of killip class values, 1 to 4

heart.rate

a numeric vector of heart rate continuous values

systolic.bp

a numeric vector of systolic blood pressure continuous values

aspirin

a binary numeric vector, 1 = yes and 0 = no

number.of.episodes.24h

a numeric vector of number of angina episodes in 24 hours

previous.pci

a binary numeric vector, 1 = yes and 0 = no

previous.cabg

a binary numeric vector, 1 = yes and 0 = no

total.chol

a numeric vector of total cholesterol values, in mmol/L

total.hdl

a numeric vector of total high density lipoprotein HDL values, in mmol/L

Value

a data frame with two extra columns including all the cardiovascular risk score calculations and their classifications

Examples

new_data_frame <- calc_scores(data = cohort_xx)
#> Error in eval(expr, envir, enclos): object 'cohort_xx' not found